falta la letra E), me da error de limite del array :'(

+++ src/Tarea3/HashDoubleE.java
@@ -31,7 +31,7 @@ public class HashDoubleE {
     }
 //-------------------------------------------------------------
 
-    public int hashFunc1(long key) {
+    public long hashFunc1(long key) {
         return Libro.trunk(key*key)%131;
     }
 //-------------------------------------------------------------
